+++ apisix/plugins/opa/init.lua
